requirements.txt
bleak
Pillow
PyYAML

# Optional: for FFT VU meter
numpy
sounddevice

# Optional: for CPU monitoring in radar animation
psutil

